drivers: pinctrl: pinctrl-rzg2l: tweak rzg2l_gpio_free api for libgpiod
The original method will revert to gpio input mode when using libgpiod.
This situation does not happen on sysfs way because sysfs does not use
free api. Compared to the stricter libgpiod, adjustments are need as
follows:

Add memory functionality for gpio setting, upon exit, the current state
will be remembered. if the state is set to output, the output level will
not change. This ensures that the gpiod utilities can be used correctly
in user-space.